+++ CHANGELOG.rst
@@ -10,7 +10,8 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
-- ``dtool put`` command
+- ``dtool add item`` command
+- ``dtool add metadata`` command
 
 
 Changed